fix #2 add pyproject.toml
